update module 6
