--- Module6/docker-compose.yml
@@ -1,40 +1,0 @@
-version: '3.7'
-services:
-  postgres:
-    image: ankane/pgvector:v0.5.1
-    restart: always
-    environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-      - POSTGRES_DB=vector_store
-      - PGUSER=postgres
-      - PGPASSWORD=postgres
-      - POSTGRES_HOST_AUTH_METHOD=trust
-    logging:
-      options:
-        max-size: 10m
-        max-file: "3"
-    ports:
-      - '5432:5432'
-    networks:
-      - postgres-network
-    healthcheck:
-      test: "pg_isready -U postgres -d vector_store"
-      interval: 2s
-      timeout: 20s
-      retries: 10
-  pgadmin:
-    container_name: pgadmin_container
-    image: dpage/pgadmin4
-    environment:
-      PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL:-pgadmin4@pgadmin.org}
-      PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD:-admin}
-    volumes:
-      - ./servers.json:/pgadmin4/servers.json
-    ports:
-      - "${PGADMIN_PORT:-5050}:80"
-    networks:
-      - postgres-network
-networks:
-  postgres-network:
-    driver: bridge
