--- Module6/src/main/java/com/epam/training/gen/ai/model/ChatRequest.java
@@ -1,4 +1,0 @@
-package com.epam.training.gen.ai.model;
-
-public record ChatRequest(String question) {
-}
