--- Module6/src/test/resources/chat.http
@@ -1,20 +1,0 @@
-### Copilot in Power Automate
-# Expected: 93% of Fortune 500 companies use Power Automate
-
-POST {{host}}/api/v1/chat
-Content-Type: application/json
-
-{
-  "question": "How many of Fortune 500 companies use Power Automate?"
-}
-
-
-### Clean Design Self-Assessment
-# Expected: Your total score is 86%
-
-POST {{host}}/api/v1/chat
-Content-Type: application/json
-
-{
-  "question": "What is my total score?"
-}
