+++ lab6/src/main/java/com/epam/training/gen/ai/model/ChatRequest.java
@@ -1,7 +1,4 @@
 package com.epam.training.gen.ai.model;
 
-import jakarta.validation.constraints.NotBlank;
-
-public record ChatRequest(
-        @NotBlank(message = "Query cannot be blank") String query) {
+public record ChatRequest(String question) {
 }
