+++ lab6/src/main/java/com/epam/training/gen/ai/model/ChatResponse.java
@@ -1,4 +1,4 @@
 package com.epam.training.gen.ai.model;
 
-public record ChatResponse(String output) {
+public record ChatResponse(String answer) {
 }
