+++ lab6/src/main/java/com/epam/training/gen/ai/service/StoreService.java
