+++ lab6/src/main/resources/application.properties
@@ -1,17 +1,22 @@
 spring.application.name=lab6
-spring.ai.azure.openai.apiKey=${AZURE_OPENAI_KEY}
-spring.ai.azure.openai.endpoint=${AZURE_OPENAI_ENDPOINT}
+server.port=8043
+logging.level.org.springframework.ai.azure.openai=DEBUG
+spring.servlet.multipart.max-file-size=200MB
+spring.servlet.multipart.max-request-size=200MB
 
-spring.threads.virtual.enabled=true
+#Spring AI
 spring.ai.openai.embedding.options.model=text-embedding-ada-002
+spring.datasource.password=postgres
+spring.datasource.username=postgres
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost/vector_store
 spring.ai.vectorstore.pgvector.index-type=none
 spring.ai.vectorstore.pgvector.distance-type=COSINE_DISTANCE
+spring.ai.azure.openai.apiKey=${AZURE_OPENAI_KEY}
+spring.ai.azure.openai.endpoint=https://ai-proxy.lab.epam.com
+spring.ai.azure.openai.chat.options.deployment-name=gpt-35-turbo
 
-spring.datasource.url=jdbc:postgresql://localhost/vector_store
-#spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.port=5432
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.datasource.database=vector_store
-spring.datasource.table=embeddings
-spring.datasource.dimension=1536
+#SK
+client.azureopenai.key=${AZURE_OPENAI_KEY}
+client.azureopenai.endpoint=${AZURE_OPENAI_ENDPOINT}
+client.azureopenai.deployment-name=gpt-35-turbo
