--- lab6/src/test/java/com/epam/training/gen/ai/RagApplicationTests.java
@@ -1,13 +1,0 @@
-package com.epam.training.gen.ai;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RagApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
